Add /* default */ to indicate intentional package-private access
Package-private (default) access (with no explicit access modifier)
is used in two circumstances:

1. Someone made a mistake and meant to include another modifier,
   often private. For some reason, annotations (such as @Mock) seem
   to encourage this forgetfulness.

2. Someone actually wanted to use package-private.

Fix cases of 1 and add the comment /* default */ to cases of 2. The
/* default */ comment is recognised by PMD’s
CommentDefaultAccessModifier rule, which will complain if package-
private access is used without it. This provides a nice balance of
allowing us the freedom to use package-private access if we wish
but guarding against accidental usage.